rename abstract to util enums
